End user is able to create authors.
