+++ app/controllers/books/books_library_controller.rb
@@ -1,0 +1,6 @@
+module Books
+	class BooksLibraryController < ApplicationController
+		def homepage
+		end
+	end
+end
