+++ app/helpers/books_library_helper.rb
@@ -1,0 +1,2 @@
+module BooksLibraryHelper
+end
