--- app/models/author.rb
@@ -1,2 +1,0 @@
-class Author < ActiveRecord::Base
-end
