--- app/views/authors/_form.html.slim
@@ -1,9 +1,0 @@
-= simple_form_for(@author) do |f|
-  = f.error_notification
-
-  .form-inputs
-    = f.input :first_name
-    = f.input :last_name
-
-  .form-actions
-    = f.button :submit
