--- app/views/authors/edit.html.slim
@@ -1,8 +1,0 @@
-h1 Editing author
-
-== render 'form'
-
-= link_to 'Show', @author
-'|
-= link_to 'Back', authors_path
-
