--- app/views/authors/index.html.slim
@@ -1,23 +1,0 @@
-h1 Listing authors
-
-table
-  thead
-    tr
-      th First name
-      th Last name
-      th
-      th
-      th
-
-  tbody
-    - @authors.each do |author|
-      tr
-        td = author.first_name
-        td = author.last_name
-        td = link_to 'Show', author
-        td = link_to 'Edit', edit_author_path(author)
-        td = link_to 'Destroy', author, data: { confirm: 'Are you sure?' }, method: :delete
-
-br
-
-= link_to 'New Author', new_author_path
