--- app/views/authors/new.html.slim
@@ -1,5 +1,0 @@
-h1 New author
-
-== render 'form'
-
-= link_to 'Back', authors_path
