+++ app/views/books/authors/_form.html.slim
@@ -1,0 +1,17 @@
+= link_to books_authors_path, class: ['btn', 'btn-success'] do
+	=> icon :users
+	= t('views.authors.list')
+
+= simple_form_for [:books, @author] do |f|
+	header
+		legend
+			h3 
+	.panel-body
+		.form-inputs
+			= f.input :first_name
+			= f.input :last_name
+		.form-actions
+			= f.button :submit
+
+
+
