+++ app/views/books/authors/edit.html.slim
@@ -1,0 +1,6 @@
+- content_for(:title) { t 'views.navigation.edit' } 
+
+section.container
+  .row
+	    .centered_form
+			      = render partial: 'form'
