+++ app/views/books/authors/index.json.jbuilder
