+++ app/views/books/authors/new.html.slim
@@ -1,0 +1,6 @@
+- content_for(:title) { t 'views.authors.new_author' } 
+
+section.container
+  .row
+	    .centered_form
+			      = render partial: 'form'
