+++ app/views/books/authors/show.html.slim
@@ -7,6 +7,6 @@ p
   strong Last name:
   = @author.last_name
 
-= link_to 'Edit', edit_author_path(@author)
+= link_to 'Edit', edit_books_author_path(@author)
 '|
-= link_to 'Back', authors_path
+= link_to 'Back', books_authors_path
