+++ app/views/books/authors/show.json.jbuilder
