+++ app/views/books/books_library/homepage.html.slim
@@ -1,0 +1,5 @@
+ul
+	<!-- Authors -->
+	li class="item"
+		= link_to books_authors_path do
+			= icon 'users', t('views.authors.list')
