+++ app/views/layouts/_navigation.html.slim
@@ -21,6 +21,12 @@ nav class="navbar navbar-static-top #{content_for?(:nav_color) ? yield(:nav_colo
         li class="#{'active' if current_page?(users_present_path) || current_page?(root_path)}"
           = link_to users_present_path do
             = icon 'street-view', t('views.navigation.presence')
+
+				<!-- Libary -->
+        li class="#{'active' if current_page?(books_library_homepage_path)}"
+					= link_to books_library_homepage_path do
+						= icon 'books-library', t('views.library')
+
         - if user_signed_in? and current_user.can_update?(Lights::Policy)
           li class="#{'active' if current_page?(lights_status_path)}"
             = link_to lights_status_path do
