+++ spec/controllers/books_library_controller_spec.rb
@@ -1,0 +1,12 @@
+require 'rails_helper'
+
+RSpec.describe BooksLibraryController, type: :controller do
+
+  describe "GET #homepage" do
+    it "returns http success" do
+      get :homepage
+      expect(response).to have_http_status(:success)
+    end
+  end
+
+end
